Enable xattr tests
Updated the xattr_common.ksh helper functions to use the attr
command on Linux to manipulate xattrs.  Added an xattr.cfg file
and reworked the user/group functionality to be consist with
the existing delegate test cases.  The intent of each test
case was preserved.

* xattr_001_pos, xattr_002_neg - Updated to verity xattr=on
  and xattr=sa sytle xattrs.

* xattr_003_neg - Use user_run helper instead of su.

* xattr_004_pos - Updated to work with ext2 xattrs.

* xattr_007_neg - Updated to use attr instead of runat.

* xattr_008_pos, xattr_009_neg8_pos, xattr_010_neg -
  Test cases disables since they aren't applicable to Linux.

* xattr_011_pos - Updated to expected behavior from GNU
  versions of the tested utilities.

* xattr_012_pos - Updated to use xattrtest to create many
  small xattrs instead of a single large one.

* xattr_013_pos - Updated to use attr instead of runat.

Closes #6128

+++ tests/zfs-tests/tests/functional/xattr/xattr_013_pos.ksh
@@ -65,13 +65,25 @@ log_must umount $TESTDIR
 log_must zfs mount -o noxattr $TESTPOOL/$TESTFS
 
 # check that we can't perform xattr operations
-log_mustnot eval "runat $TESTDIR/myfile.$$ cat passwd > /dev/null 2>&1"
-log_mustnot eval "runat $TESTDIR/myfile.$$ rm passwd > /dev/null 2>&1"
-log_mustnot eval "runat $TESTDIR/myfile.$$ cp /etc/passwd . > /dev/null 2>&1"
+if is_linux; then
+	log_mustnot attr -q -g passwd $TESTDIR/myfile.$$
+	log_mustnot attr -q -r passwd $TESTDIR/myfile.$$
+	log_mustnot attr -q -s passwd $TESTDIR/myfile.$$ </etc/passwd
 
-log_must touch $TESTDIR/new.$$
-log_mustnot eval "runat $TESTDIR/new.$$ cp /etc/passwd . > /dev/null 2>&1"
-log_mustnot eval "runat $TESTDIR/new.$$ rm passwd > /dev/null 2>&1"
+	log_must touch $TESTDIR/new.$$
+	log_mustnot attr -q -s passwd $TESTDIR/new.$$ </etc/passwd
+	log_mustnot attr -q -r passwd $TESTDIR/new.$$
+else
+	log_mustnot eval "runat $TESTDIR/myfile.$$ cat passwd > /dev/null 2>&1"
+	log_mustnot eval "runat $TESTDIR/myfile.$$ rm passwd > /dev/null 2>&1"
+	log_mustnot eval "runat $TESTDIR/myfile.$$ cp /etc/passwd . \
+	    > /dev/null 2>&1"
+
+	log_must touch $TESTDIR/new.$$
+	log_mustnot eval "runat $TESTDIR/new.$$ cp /etc/passwd . \
+	    > /dev/null 2>&1"
+	log_mustnot eval "runat $TESTDIR/new.$$ rm passwd > /dev/null 2>&1"
+fi
 
 # now mount the filesystem again as normal
 log_must umount $TESTDIR
@@ -82,7 +94,11 @@ verify_xattr $TESTDIR/myfile.$$ passwd /etc/passwd
 
 # there should be no xattr on the file we created while the fs was mounted
 # -o noxattr
-log_mustnot eval "runat $TESTDIR/new.$$ cat passwd > /dev/null 2>&1"
+if is_linux; then
+	log_mustnot attr -q -g passwd $TESTDIR/new.$$
+else
+	log_mustnot eval "runat $TESTDIR/new.$$ cat passwd > /dev/null 2>&1"
+fi
 create_xattr $TESTDIR/new.$$ passwd /etc/passwd
 
 log_pass "The noxattr mount option functions as expected"
